Restructure Consumption
- There's a race with the sarama client where it may not have received
metadata from the topic by the time it's asking for partition
information. As a result, you may get panics with "broker not connected"
which just means that we should wait a bit longer for the connection to
establish and the metadata to be received.  This implements a
retry+backoff policy that checks for the specific `sarama.ErrNotConnected`
error and reattempts to get the metadata until the retry threshold has
been exhausted.
- Consumption uses multiple goroutines but (partly) just writes to
stderr/stdout without synchronization.  This changes the consumption to
use a channel so that all writes to the console are serialized without
needing to use a lock.  This creates consistent output without messages
being interleaved.
- We might want to use sarama's newer ConsumerGroup interface which
would eliminate the need for multiple go routines, but there may be some
other information (like offsets, etc.) that's not available in that API
which would make it undesireable for this purpose.