cmd/consume: improve skip and tail flags
This changes the flags for controlling starting
position to use --skip to mean "start at min + X"
and --tail to mean "start at max -X".
It also adds --skips and --takes parameters to
allow partition-specific offset setting, and
adds a --count/--counts pair with the same conventions
for limiting the number of messages consumed.